+++ public/js/src/route/index.js
@@ -1,5 +1,8 @@
 import mapUtils from '../utils/mapUtils'
 import map from '../map/index'
+import utils from '../utils/utils';
+import graphUtils from '../utils/graphUtils';
+import { UPDATE_FREQ, PAGE_SIZE, RATION, run } from '../utils/globals';
 
 /** All load functions */
 var loadRoutes = function () {
@@ -29,4 +32,68 @@ var loadRoutes = function () {
 
 };
 
-export default { loadRoutes }
+var createRouteRecord = function (htmlTable, index, route) {
+    htmlTable += '<tr>';
+    htmlTable += '<th scope="row">' + (index + 1) + '</th>';
+    htmlTable += '<td><a data-toggle="modal" data-route-name="' + route.name + '" data-target="#show-route-modal" href="/routes/' + route.name + '">' + route.name + '</a></td>';
+    htmlTable += '<td>' + parseInt(route.meters) + 'm</td>';
+    htmlTable += '<td>' + route.country + '</td>';
+    htmlTable += '<td>'
+    if (route.permanent !== true) {
+        htmlTable += '<a class="edit-route" href="#" data-id="' + route.name + '"><i class="material-icons">create</i></a><a class="del-route" href="#" data-id="' + route.name + '"><i aria-hidden="true" title="Delete route" class="material-icons">delete</i></a>' + '</td>';
+    }
+    htmlTable += '</tr>';
+    return htmlTable;
+};
+
+function loadRoute(mainIndex) {
+    $('#load').load('/route', function () {
+        $(this).find('#routes-t').each(function () {
+            loadRouteTable(0);
+            var pag = $('#routes-table').find('.page');
+            createRouteNavPage(pag[0], mainIndex);
+        });
+    });
+}
+
+function createRouteNavPage(page, index) {
+    var htmlElement = $('<ul id="route-page" data-index="' + index + '"></ul>').addClass("pagination pagination-lg");
+    $.get("routes/size", function (data) {
+        var size = parseInt(parseInt(data) / PAGE_SIZE) + 1;
+        var prevDisabled = (index === 0 ? 'disabled' : '');
+        var nextDisabled = (index === size - 1 ? 'disabled' : '');
+        var prev = $('<li class="page-item ' + prevDisabled + '"></li>').append('<a class="page-link" href="#" data-next="-1" tabindex="-1">Previous</a>');
+        var next = $('<li class="page-item ' + nextDisabled + '"></li>').append('<a class="page-link" data-next="1" href="#">Next</a>');
+
+        htmlElement.append(prev);
+        for (var i = 0; i < size; i++) {
+            var active = '';
+            if (index === i) {
+                active = "active";
+            }
+            var item = $('<li class="page-item ' + active + '"><a class="page-link" data-index="' + i + '" href="#">' + (i + 1) + '</a></li>');
+            htmlElement.append(item);
+        }
+        htmlElement.append(next);
+        $(page).html(htmlElement);
+    });
+}
+
+function loadRouteTable(mainIndex) {
+    var start = mainIndex * PAGE_SIZE, stop = (((mainIndex + 1) * PAGE_SIZE)) - 1;
+    $.get('/routes/' + start + '/' + stop, function (data) {
+        var htmlTable = '';
+        var index = 0;
+        data.forEach(function (route) {
+            htmlTable = createRouteRecord(htmlTable, index + (mainIndex * PAGE_SIZE), route);
+            index++;
+        });
+
+        $('#routes-table-body').html(htmlTable);
+        $('#add-route-modal').on('hidden.bs.modal', function (e) {
+            loadRoute(0);
+        })
+    });
+}
+
+export default { loadRoutes, createRouteRecord, loadRoute, loadRouteTable, createRouteNavPage }
