+++ public/js/src/utils/globals.js
@@ -1,5 +1,7 @@
 export const RATION = (100 / 4.805) * 6;
 export const PAGE_SIZE = 10;
 export const UPDATE_FREQ = 1000;
+export var run = false;
 
-export default { RATION, PAGE_SIZE, UPDATE_FREQ }
+
+export default { RATION, PAGE_SIZE, UPDATE_FREQ, run }
