+++ public/js/src/utils/graphUtils.js
@@ -1,0 +1,164 @@
+import { UPDATE_FREQ, PAGE_SIZE, RATION, run } from '../utils/globals';
+import utils from '../utils/utils';
+
+function addGraph(time, hr, start, strokes) {
+    var speed = [];
+    var watt = [];
+    var stroke = [];
+    var strokeConter = 1;
+    for (var i = 1; i < time.length; i++) {
+        var timeVal = parseInt(time[i]);
+        var strokeTime = parseInt(strokes[strokeConter]);
+        var sec = ((timeVal - start) / 1000);
+        var lenght = (RATION / 100);
+        speed.push(((lenght / sec)) * 3.6);
+        var wattValue = utils.calcWatt(sec / lenght);
+        watt.push(wattValue);
+        stroke.push(1000*60 / (strokeTime - parseInt(strokes[strokeConter-1])));
+        start = parseInt(time[i]);
+        if (timeVal > strokeTime) {
+            strokeConter++;
+        }
+    }
+
+    //Remove ever second element
+    var speedMerged = [];
+    var hrMerged = [];
+    var wattMerged = [];
+    var strokeMerged = [];
+    var labelsMerged = [];
+    var mergeSize = 10;
+    if (time.length > 1000) {
+        mergeSize = 20;
+    }
+
+    while (time.length) {
+        var a = time.splice(0, mergeSize);
+        var timeV = parseInt(a.reduce(function (a, b) { return a + b; }) / a.length);
+        labelsMerged.push(new Date(timeV).toISOString().substr(new Date(timeV).toISOString().lastIndexOf('T') + 1, 8));
+        if (hr) {
+            var h = hr.splice(0, mergeSize);
+            if (h.length > 0) {
+                hrMerged.push(parseInt(h.reduce(function (a, b) { return a + b; }) / h.length));
+            }
+        }
+        if (stroke) {
+            var h = stroke.splice(0, mergeSize);
+            if (h.length > 0) {
+                strokeMerged.push(Math.round(parseFloat(h.reduce(function (a, b) { return a + b; }) / h.length) * 10) / 10);
+            }
+        }
+        if (speed) {
+            var s = speed.splice(0, mergeSize);
+            var w = watt.splice(0, mergeSize);
+            if (s.length > 0) {
+                speedMerged.push(Math.round(parseFloat(s.reduce(function (a, b) { return a + b; }) / s.length) * 10) / 10);
+            }
+            if (w.length > 0) {
+                wattMerged.push(Math.round(parseFloat(w.reduce(function (a, b) { return a + b; }) / w.length) * 10) / 10);
+            }
+        }
+    }
+
+    var ctx = $('#hr-graph');
+    var lineChartData = {
+        labels: labelsMerged,
+        datasets: [{
+            label: 'Heart rate (bpm)',
+            borderColor: '#dc3545',
+            backgroundColor: '#dc3545',
+            fill: false,
+            data: hrMerged,
+            // cubicInterpolationMode: 'monotone',
+            yAxisID: 'y-axis-1',
+        }, {
+            label: 'Speed (km/t)',
+            borderColor: '#007bff',
+            backgroundColor: '#007bff',
+            fill: false,
+            data: speedMerged,
+            //cubicInterpolationMode: 'monotone',
+            yAxisID: 'y-axis-2'
+        },
+            {
+                label: 'Watt',
+                borderColor: '#4bc0c0',
+                backgroundColor: '#4bc0c0',
+                fill: false,
+                data: wattMerged,
+                lineTension: 0,
+                //cubicInterpolationMode: 'monotone',
+                yAxisID: 'y-axis-3'
+            },
+            {
+                label: 'Stroke rate (spm)',
+                borderColor: '#9966FF',
+                backgroundColor: '#9966FF',
+                fill: false,
+                data: strokeMerged,
+                yAxisID: 'y-axis-4'
+            }]
+    };
+    var myLineChart = Chart.Line(ctx, {
+        data: lineChartData,
+        options: {
+            responsive: true,
+            hoverMode: 'index',
+            stacked: false,
+            scales: {
+                yAxes: [{
+                    type: 'linear', // only linear but allow scale type registration. This allows extensions to exist solely for log scale for instance
+                    display: true,
+                    position: 'left',
+                    id: 'y-axis-1',
+                    ticks: {
+                        suggestedMin: 30,
+                        min: 0,
+                        stepSize: 5
+                    }
+                }, {
+                    type: 'linear', // only linear but allow scale type registration. This allows extensions to exist solely for log scale for instance
+                    display: true,
+                    position: 'right',
+                    id: 'y-axis-2',
+                    ticks: {
+                        stepSize: 2
+                    },
+                    // grid line settings
+                    gridLines: {
+                        drawOnChartArea: false // only want the grid lines for one axis to show up
+                    }
+                },
+                    {
+                        type: 'linear', // only linear but allow scale type registration. This allows extensions to exist solely for log scale for instance
+                        display: true,
+                        position: 'right',
+                        id: 'y-axis-3',
+                        ticks: {
+                            stepSize: 25
+                        },
+                        // grid line settings
+                        gridLines: {
+                            drawOnChartArea: false // only want the grid lines for one axis to show up
+                        }
+                    },
+                    {
+                        type: 'linear', // only linear but allow scale type registration. This allows extensions to exist solely for log scale for instance
+                        display: true,
+                        position: 'right',
+                        id: 'y-axis-4',
+                        ticks: {
+                            stepSize: 2,
+                            suggestedMin: 10,
+                        },
+                        // grid line settings
+                        gridLines: {
+                            drawOnChartArea: false // only want the grid lines for one axis to show up
+                        }
+                    }]
+            }
+        }
+    });
+}
+
+export default { addGraph }
